requirements.txt
numpy==1.19.5
pandas==1.2.2
PyWavelets==1.1.1
Pillow==8.3.1
scikit_learn==0.24.2
